add copies admin endpoints

+++ app/Http/Controllers/API/V1/CopyController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\V1\CopyCollection;
 use App\Http\Resources\V1\CopyResource;
 use App\Models\Copy;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class CopyController extends Controller
@@ -23,6 +24,7 @@ class CopyController extends Controller
         return new CopyCollection(Copy::all()->where('user_id', '=', $userId));
     }
 
+
     /**
      * Store a newly created resource in storage.
      *
@@ -32,7 +34,6 @@ class CopyController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            'userId'        => ['integer', 'required'],
             'gameId'        => ['integer','required'],
             'platformId'    => ['integer','required'],
             'rating'        => ['integer', 'min:1', 'max:5'],
@@ -40,8 +41,10 @@ class CopyController extends Controller
 
         ]);
 
+        $userId = auth()->user()->id;
+
         $copy = Copy::create([
-            'user_id'       => $request->input('userId'),
+            'user_id'       => $userId,
             'game_id'       => $request->input('gameId'),
             'platform_id'   => $request->input('platformId'),
             'rating'        => $request->input('rating'),
@@ -61,7 +64,21 @@ class CopyController extends Controller
      */
     public function show(Copy $copy)
     {
-        return (new CopyResource($copy))
+
+        $userId = auth()->user()->id;
+
+        $cop = new CopyResource($copy);
+
+        $copyOwnerId = $cop->user_id;
+
+        if($copyOwnerId !== $userId) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Unauthorized',
+            ], 401);
+        }
+
+        return ($cop)
             ->response()
             ->setStatusCode((200));
     }
@@ -75,8 +92,21 @@ class CopyController extends Controller
      */
     public function update(Request $request, Copy $copy)
     {
+
+        $userId = auth()->user()->id;
+
+        $cop = new CopyResource($copy);
+
+        $copyOwnerId = $cop->user_id;
+
+        if($copyOwnerId !== $userId) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Unauthorized',
+            ], 401);
+        }
+
         $this->validate($request, [
-            'userId'        => ['integer'],
             'gameId'        => ['integer'],
             'platformId'    => ['integer'],
             'rating'        => ['integer', 'min:1', 'max:5'],
@@ -84,14 +114,13 @@ class CopyController extends Controller
         ]);
 
         $copy->update([
-            'user_id'       => $request->input('userId') ?? $copy->user_id,
             'game_id'       => $request->input('gameId') ?? $copy->game_id,
             'platform_id'   => $request->input('platformId') ?? $copy->platform_id,
             'rating'        => $request->input('rating') ?? $copy->rating,
             'completed'     => $request->input('completed') ?? $copy->completed,
         ]);
 
-        return (new CopyResource($copy))
+        return ($cop)
             ->response()
             ->setStatusCode(200);
     }
@@ -104,8 +133,52 @@ class CopyController extends Controller
      */
     public function destroy(Copy $copy)
     {
+
+        $userId = auth()->user()->id;
+
+        $cop = new CopyResource($copy);
+
+        $copyOwnerId = $cop->user_id;
+
+        if($copyOwnerId !== $userId) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Unauthorized',
+            ], 401);
+        }
+
         $copy->delete();
 
         return response()->json(null, 204);
+    }
+
+
+
+    public function indexUserCopies(User $user)
+    {
+        return new CopyCollection(Copy::all()->where('user_id', '=', $user->id));
+    }
+
+    public function storeUserCopy(Request $request, User $user)
+    {
+        $this->validate($request, [
+            'gameId'        => ['integer','required'],
+            'platformId'    => ['integer','required'],
+            'rating'        => ['integer', 'min:1', 'max:5'],
+            'completed'     => ['boolean'],
+
+        ]);
+
+        $copy = Copy::create([
+            'user_id'       => $user->id,
+            'game_id'       => $request->input('gameId'),
+            'platform_id'   => $request->input('platformId'),
+            'rating'        => $request->input('rating'),
+            'completed'     => $request->input('completed'),
+        ]);
+
+        return (new CopyResource($copy))
+            ->response()
+            ->setStatusCode(201);
     }
 }
